refactor: 이미지 크롭 기능에 맞게 로직 분리
