+++ src/app/_components/atoms/ImageCropperHeader.tsx
@@ -1,30 +1,55 @@
-import React, { forwardRef, KeyboardEvent } from 'react';
+import React, { KeyboardEvent, useCallback, useEffect, useRef } from 'react';
 import BackIcon from '@/assets/icons/BackIcon';
 import CheckIcon from '@/assets/icons/CheckIcon';
+import { initialImage, useUploadImage } from '@/store/uploadImage';
+import { useShallow } from 'zustand/react/shallow';
 
 type CropImgHeaderProps = {
-  handleCancelCrop: () => void;
-  handleKeyDownCancelCrop: (e: KeyboardEvent<HTMLButtonElement>) => void;
   handleImgCrop: () => void;
+  movePage: () => void;
   handleKeyDownImgCrop: (e: KeyboardEvent<HTMLButtonElement>) => void;
 };
 
-const ImageCropperHeader = forwardRef<HTMLButtonElement, CropImgHeaderProps>(
-  (
-    {
-      handleCancelCrop,
-      handleKeyDownCancelCrop,
-      handleImgCrop,
-      handleKeyDownImgCrop,
+export default function ImageCropperHeader({
+  handleImgCrop,
+  handleKeyDownImgCrop,
+  movePage,
+}: CropImgHeaderProps) {
+  const cancelButtonRef = useRef<HTMLButtonElement>(null);
+  const { resetUploadImageState, setUploadImage } = useUploadImage(
+    useShallow((state) => ({
+      uploadImage: state.uploadImage,
+      setUploadImage: state.setUploadImage,
+      setCropedPreview: state.setCropedPreview,
+      resetUploadImageState: state.resetUploadImageState,
+    })),
+  );
+
+  useEffect(() => {
+    cancelButtonRef.current?.focus();
+  }, []);
+
+  const handleCancelCrop = useCallback(() => {
+    setUploadImage(initialImage);
+    movePage();
+  }, []);
+
+  const handleKeyDownCancelCrop = useCallback(
+    (e: KeyboardEvent<HTMLButtonElement>) => {
+      if (e.key === 'Enter') {
+        resetUploadImageState();
+      }
     },
-    ref,
-  ) => (
+    [handleCancelCrop],
+  );
+
+  return (
     <header
       aria-label="이미지 자르기 작업 메뉴"
       className="dark:text-dark-line-light text-dark-bg-dark font-semibold flex flex-row justify-between items-center p-10"
     >
       <button
-        ref={ref}
+        ref={cancelButtonRef}
         type="button"
         aria-label="이미지 변경 취소하기"
         className="font-normal flex items-center gap-x-5 flex-grow basis-1/3 focus:focus-sr"
@@ -47,7 +72,7 @@ const ImageCropperHeader = forwardRef<HTMLButtonElement, CropImgHeaderProps>(
         <CheckIcon className="w-25 h-25" />
       </button>
     </header>
-  ),
-);
+  );
+}
 
-export default React.memo(ImageCropperHeader);
+// export default React.memo(ImageCropperHeader);
